prosta walidacja

+++ app/Http/Controllers/MailingController.php
@@ -3,6 +3,7 @@
 namespace Coyote\Http\Controllers;
 
 use Coyote\Repositories\Contracts\MailingRepositoryInterface as MailingRepository;
+use Ramsey\Uuid\Uuid;
 
 class MailingController extends Controller
 {
@@ -13,6 +14,15 @@ class MailingController extends Controller
      */
     public function unsubscribe($uuid, MailingRepository $mailing)
     {
+        if (!Uuid::isValid($uuid)) {
+            return redirect()
+                ->to('/')
+                ->with(
+                    'error',
+                    sprintf('Ups. Link wygląda na nieprawidłowy. Napisz do nas na %s, a usuniemy Twój adres e-mail z listy mailingowej.', config('mail.from.address'))
+                );
+        }
+
         /** @var \Coyote\Mailing $result */
         $result = $mailing->findOrFail($uuid);
         $result->delete();
